Revert "chore: split LocationTracker into enrichment and sync concerns, add NOT_DETERMINED auth status"
This reverts commit e480d68cdbdcaaf9ea35333aee4c53dfa93e4d57.
