--- location/src/main/kotlin/io/customer/location/LocationCoordinates.kt
@@ -1,9 +1,0 @@
-package io.customer.location
-
-/**
- * Internal location coordinate holder.
- */
-internal data class LocationCoordinates(
-    val latitude: Double,
-    val longitude: Double
-)
